Combine CAS 22 reactor plant equipment total into a single shared function
